part3/requirements.txt
fastapi==0.110.0
pydantic==2.6.4
aiofiles==23.2.1
boto3==1.34.130
requests==2.31.0
Pillow==10.2.0
moviepy==1.0.3
mangum==0.17.0
loguru==0.7.2
python-multipart==0.0.9
uvicorn==0.28.1
